Refactor common code into hid-gcore. Remove hid-ginput
lg4l modules were created by copy-paste from an initial implementation.
This means most fixes have to be applied into all modules, which creates
a drag. Time for a refactoring :)

Removed hid-ginput and the keyboard-switching feature from the kernel
driver code for better conformance to the linux kernel philosophy of not
implementing policies, but features. The keymap switch can just as
easily be implemented in an userspace daemon using ioctls, or as a
higher-level policy module if absolutely needed.